requirements.txt
matplotlib
numpy
torch
torchvision
Pillow
tqdm
GPUtil
IPython
